+++ Components/IRacing/Plugins/Trackers/Trackers.cs
@@ -29,6 +29,7 @@ namespace Slipstream.Components.IRacing.Plugins.Trackers
             DataTrackers.Add(new IRacingSessionTracker(eventBus, eventFactory));
             DataTrackers.Add(new CarPositionTracker(eventBus, eventFactory));
             DataTrackers.Add(new TowTracker(eventBus, eventFactory));
+            DataTrackers.Add(new TrackPositionTracker(eventBus, eventFactory));
         }
 
         public void Handle(GameState.IState currentState)